pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "github-manager-mcp"
version = "0.1.0"
description = "FastMCP server for comprehensive GitHub management - repositories, automation, workspace, and backup"
authors = [
    {name = "KoreanLeeChangHyun", email = "m.leechanghyun@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
keywords = ["github", "mcp", "fastmcp", "automation", "backup", "workspace"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "fastmcp>=0.1.0",
    "PyGithub>=2.1.1",
    "pydantic>=2.5.0",
    "python-dotenv>=1.0.0",
    "gitpython>=3.1.40",
    "pyyaml>=6.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-cov>=4.1.0",
    "black>=23.12.0",
    "ruff>=0.1.8",
    "mypy>=1.7.1",
]

[project.scripts]
github-manager-mcp = "github_manager.server:main"

[project.urls]
Homepage = "https://github.com/KoreanLeeChangHyun/github-manager"
Repository = "https://github.com/KoreanLeeChangHyun/github-manager"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ['py39', 'py310', 'py311', 'py312']

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "--cov=src/github_manager --cov-report=term-missing"
